fix(db): model int64in/int64out DBF_INT64 and aSub DBF_LONG fields as integers so string puts parse via epicsParseInt64/Long
Cause B of the put-conversion divergence family. int64in/int64out modeled
their DBF_INT64 operator/drive fields (HOPR/LOPR/DRVH/DRVL/HYST/IVOV/ADEL/MDEL)
as f64, and aSub modeled its DBF_LONG VAL (the subroutine return status) as f64.
The put target is resolved from the field's STORED type (get_field's variant),
so those fields keyed the string->native put parse on DBF_DOUBLE and
c_parse::put_string ran the float row -- accepting 9.22e18 (2^63) where C's
putStringInt64 -> epicsParseInt64 refuses it. 15 oracle put-accept defects.

Structural fix: store/serve the fields in their C-declared integer variant, so
get_field reports Int64/Long, the put target becomes Int64/Long, and the
existing single owner c_parse::put_string parses via the integer row (rejects
out-of-range, keeps strtol trailing-text tolerance). No range logic is
reimplemented. Served over CA unchanged: EpicsValue::Int64 maps to DBR_DOUBLE
on the wire (the record already served int64 as double, as C does), Long stays
DBR_LONG. int64out::process's drive clamp is now a direct i64 compare, so the
double->i64 saturating rounding helper is gone.

aSub VAL put_field stays lenient (any numeric -> i32, C's `prec->val = status`
assignment) because it is a status field the framework and subroutine closures
assign directly; the parse rejection lives in coerce_put_value ahead of it, so
leniency here does not reopen the divergence.

Scope: only the WRITABLE DBF_INT64/DBF_LONG record-owned fields. Left as f64:
LALM/ALST/MLST (DBF_INT64 but special(SPC_NOMOD) read-only -- no client put
reaches the parse, oracle agrees); HIHI/HIGH/LOW/LOLO (DBF_INT64 but routed to
the shared common.analog_alarm, which also serves ao/ai's DBF_DOUBLE limits);
AFTC/AFVL/SDLY (genuinely DBF_DOUBLE).

Scoped put oracle (int64in,int64out,aSub): defect 41->26, 0 errored, 0 new;
the 15 closed are exactly the int64/aSub put-accept cases. Remaining 26 are
out of Cause B: aSub SNAM (Cause C, 5) and the enum-over/negative-ordinal
menu-index family (21).

Regression test int64_field_put_is_integer_parsed.rs: over-i64/i32 refused +
field unchanged; in-range and band edges accepted and stored as Int64/Long;
trailing-text/fraction truncate not reject; CA wire type unchanged. Existing
aSub VAL assertions updated Double->Long to match the served DBF_LONG.

+++ crates/epics-base-rs/src/server/record/record_instance.rs
@@ -3291,9 +3291,11 @@ impl RecordInstance {
         // closure may have written to VAL. `sub` does NOT do this — its VAL
         // is the value the subroutine computed.
         if self.record.record_type() == "aSub" {
+            // aSub VAL is DBF_LONG (epicsInt32); the do_sub status is a C `long`
+            // truncated into it (`prec->val = status`).
             let _ = self
                 .record
-                .put_field("VAL", EpicsValue::Double(status as f64));
+                .put_field("VAL", EpicsValue::Long(status as i32));
         }
 
         // A negative status raises SOFT_ALARM at the record's BRSV severity
